Moved parser logic into its own file
