+++ PowerballApi/Api.csproj
@@ -91,7 +91,7 @@
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
-    <Compile Include="Helper\Parser.cs" />
+    <Compile Include="Helper\PowerballParser.cs" />
     <Compile Include="Models\PowerballSet.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
