+++ PowerballApi/Controllers/PowerballController.cs
@@ -5,7 +5,7 @@
 	using System.IO;
 	using System.Net;
 	using System.Web.Http;
-	using Microsoft.VisualBasic.FileIO;
+	using Helper;
 	using Models;
 
 	public class PowerballController : ApiController
@@ -14,41 +14,9 @@
 		public IEnumerable<PowerballSet> GetPowerballNumbers()
 		{
 			var file = GetPowerballFile();
-			var powerballResults = new List<PowerballSet>();
-
-			// TODO: Isolate parser for single responsibility
-			using (var parser = new TextFieldParser(file))
-			{
-				parser.SetDelimiters("  ");
-
-				while (!parser.EndOfData)
-				{
-					if (parser.LineNumber == 1)
-						continue;
-
-					var dataLine = parser.ReadFields();
-					if (dataLine == null)
-						continue;
-
-					var set = new PowerballSet
-					{
-						Date = dataLine[0],
-						WinNumbers =
-						{
-							[0] = int.Parse(dataLine[1]),
-							[1] = int.Parse(dataLine[2]),
-							[2] = int.Parse(dataLine[3]),
-							[3] = int.Parse(dataLine[4]),
-							[4] = int.Parse(dataLine[5]),
-							[5] = int.Parse(dataLine[6])
-						}
-					};
-
-					powerballResults.Add(set);
-				}
-			}
-
-			return powerballResults;
+			var parser = new PowerballParser();
+			var results = parser.Parse(file);
+			return results;
 		}
 
 		private static string GetPowerballFile()
