--- PowerballApi/Helper/Parser.cs
@@ -1,6 +1,0 @@
-﻿namespace PowerballApi.Helper
-{
-    public class Parser
-    {
-    }
-}
