+++ PowerballApi/Helper/PowerballParser.cs
@@ -1,0 +1,49 @@
+﻿namespace PowerballApi.Helper
+{
+	using System.Collections.Generic;
+	using Microsoft.VisualBasic.FileIO;
+	using Models;
+
+	public class PowerballParser
+	{
+		public List<PowerballSet> Parse(string file)
+		{
+			if (string.IsNullOrWhiteSpace(file))
+				return new List<PowerballSet>();
+
+			var powerballResults = new List<PowerballSet>();
+			using (var parser = new TextFieldParser(file))
+			{
+				parser.SetDelimiters("  ");
+
+				while (!parser.EndOfData)
+				{
+					if (parser.LineNumber == 1)
+						continue;
+
+					var dataLine = parser.ReadFields();
+					if (dataLine == null || dataLine.Length < 7)
+						continue;
+
+					var set = new PowerballSet
+					{
+						Date = dataLine[0],
+						WinNumbers =
+						{
+							[0] = int.Parse(dataLine[1]),
+							[1] = int.Parse(dataLine[2]),
+							[2] = int.Parse(dataLine[3]),
+							[3] = int.Parse(dataLine[4]),
+							[4] = int.Parse(dataLine[5]),
+							[5] = int.Parse(dataLine[6])
+						}
+					};
+
+					powerballResults.Add(set);
+				}
+			}
+
+			return powerballResults;
+		}
+    }
+}
